file structure merged
