--- tapir/shifts/templates/shifts/calendarview.html
@@ -1,28 +1,0 @@
-{% extends "shifts/base.html" %}
-
-{% load django_bootstrap5 %}
-{% load static %}
-{% load shifts %}
-{% load i18n %}
-{% load coop %}
-
-{% block head %}
-    {{ block.super }}
-    <link rel="stylesheet" href="{% static "shifts/css/calendar.css" %}">
-{% endblock %}
-
-
-
-{% block content %}
-    <div class="container">
-        <div class="card m-2">
-            <h5 class="card-header">
-                <span>{% translate "Shift calendar" %}</span>
-            </h5>
-            <div class="card-body">
-                <p><a href="{% static "shifts/ABCD_weeks_calendar_EN.pdf" %}">{% translate 'Download calendar as pdf' %}</a></p>
-                {{ shiftcal }}
-            </div>
-        </div>
-    </div>
-{% endblock %}
