+++ tapir/shifts/urls.py
@@ -118,9 +118,4 @@ urlpatterns = [
         views.StatisticsView.as_view(),
         name="statistics",
     ),
-    path(
-        "calendarview",
-        views.ShiftCalendarView.as_view(),
-        name="calendarview",
-    ),
 ]
