+++ tapir/shifts/views/calendars.py
@@ -17,7 +17,7 @@ from tapir.shifts.models import (
 )
 from tapir.shifts.templatetags.shifts import get_week_group
 from tapir.shifts.views.views import get_shift_slot_names, SelectedUserViewMixin
-from tapir.shifts.year_schedule import ColorHTMLCalendar
+from tapir.shifts.utils import ColorHTMLCalendar
 
 
 class ShiftCalendarBaseView(TemplateView):
@@ -140,20 +140,6 @@ class ShiftTemplateOverview(LoginRequiredMixin, SelectedUserViewMixin, TemplateV
 
 class ShiftTemplateGroupCalendar(LoginRequiredMixin, TemplateView):
     template_name = "shifts/shift_template_group_calendar.html"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data()
-        date_to_group = {}
-        today = timezone.now().date()
-        for week in range(52):
-            monday = today - timedelta(days=today.weekday()) + timedelta(weeks=week)
-            date_to_group[monday] = get_week_group(monday).name
-        context["date_to_group"] = date_to_group
-        return context
-
-
-class ShiftCalendarView(LoginRequiredMixin, TemplateView):
-    template_name = "shifts/calendarview.html"
 
     def get_context_data(self, *args, **kwargs):
         context = super().get_context_data(*args, **kwargs)
